requirements.txt
numpy==1.23.5
pandas
scikit-learn
matplotlib
seaborn
opencv-python-headless
pillow
tqdm
albumentations
imbalanced-learn